+++ docs/src/markdown/changelog.md
@@ -7,7 +7,7 @@
 - **NEW**: Include a default notification sound and use it as the default across all platforms.
 - **NEW**: Provide fallback players for Linux: `paplay`, `aplay`, and `play` (`sox`).
 - **NEW**: Provide interface for selecting custom notification sounds.
-- **FIX**: Better attempt to get appropriate creation on all Linux systems.
+- **FIX**: Better attempt to get appropriate file creation time on all Linux systems.
 - **FIX**: Better fix for time picker getting wrong background.
 - **FIX**: Ensure column sizing includes header size.
 - **FIX**: When showing timed status event for results in status bar (full path name), decrease length of time that the temporary status is shown and clear temporary status when the cursor leaves result items.